Resume Project/requirements.txt
# Core dependencies
flask==2.3.3
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1

# Document processing
PyMuPDF==1.23.8
pdfplumber==0.10.3
python-docx==1.1.0
docx2txt==0.8

# NLP and ML
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2
openai==1.3.7
sentence-transformers==2.2.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Data processing
pandas==2.1.3
numpy==1.25.2
python-multipart==0.0.6

# Report generation
reportlab==4.0.7
jinja2==3.1.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
